Rename from rails-sort to rails_sort
