+++ .ruby-gemset
@@ -1,1 +1,1 @@
-rails-sort
+rails_sort
