+++ lib/rails_sort/active_record_patch.rb
