+++ lib/rails_sort/query_string.rb
