+++ spec/lib/rails_sort/query_string_spec.rb
