+++ spec/lib/rails_sort/uri_constructor_spec.rb
