+++ spec/spec_helper.rb
@@ -2,7 +2,7 @@ require 'bundler'
 Bundler.require
 require 'rails/all'
 
-# Dir.glob("./lib/rails-sort/**/*.rb", &method(:require))
+# Dir.glob("./lib/rails_sort/**/*.rb", &method(:require))
 require './lib/rails_sort'
 Dir.glob("./lib/support/**/*.rb", &method(:require))
